Move commands from sample project to Toolkit
