+++ Microsoft.Toolkit.Uwp/Commands/DelegateCommand.cs
@@ -14,23 +14,38 @@ using System;
 using System.Diagnostics;
 using System.Windows.Input;
 
-namespace Microsoft.Toolkit.Uwp.SampleApp.Common
+namespace Microsoft.Toolkit.Uwp.Commands
 {
-    public class DelegateCommand<T> : ICommand
+    /// <summary>
+    /// Represents a command that can perform a given action.
+    /// </summary>
+    public class DelegateCommand : ICommand
     {
-        private readonly Action<T> commandExecuteAction;
+        private readonly Action _commandExecuteAction;
 
-        private readonly Func<T, bool> commandCanExecute;
+        private readonly Func<bool> _commandCanExecute;
 
-        public DelegateCommand(Action<T> executeAction, Func<T, bool> canExecute = null)
+        /// <summary>
+        /// Initializes a new instance of the <see cref="DelegateCommand"/> class.
+        /// </summary>
+        /// <param name="execute">
+        /// The action to execute when called.
+        /// </param>
+        /// <param name="canExecute">
+        /// The function to call to determine if the command can execute the action.
+        /// </param>
+        /// <exception cref="ArgumentNullException">
+        /// Thrown if the execute action is null.
+        /// </exception>
+        public DelegateCommand(Action execute, Func<bool> canExecute = null)
         {
-            if (executeAction == null)
+            if (execute == null)
             {
-                throw new ArgumentNullException(nameof(executeAction));
+                throw new ArgumentNullException(nameof(execute));
             }
 
-            commandExecuteAction = executeAction;
-            commandCanExecute = canExecute ?? (e => true);
+            _commandExecuteAction = execute;
+            _commandCanExecute = canExecute ?? (() => true);
         }
 
         /// <summary>
@@ -47,11 +62,11 @@ namespace Microsoft.Toolkit.Uwp.SampleApp.Common
         /// <returns>
         /// Returns a value indicating whether this command can be executed.
         /// </returns>
-        public bool CanExecute(object parameter)
+        public bool CanExecute(object parameter = null)
         {
             try
             {
-                return commandCanExecute(ConvertParameterValue(parameter));
+                return _commandCanExecute();
             }
             catch
             {
@@ -74,7 +89,7 @@ namespace Microsoft.Toolkit.Uwp.SampleApp.Common
 
             try
             {
-                commandExecuteAction(ConvertParameterValue(parameter));
+                _commandExecuteAction();
             }
             catch
             {
@@ -82,15 +97,12 @@ namespace Microsoft.Toolkit.Uwp.SampleApp.Common
             }
         }
 
+        /// <summary>
+        /// Raises the can execute changed.
+        /// </summary>
         public void RaiseCanExecuteChanged()
         {
             CanExecuteChanged?.Invoke(this, EventArgs.Empty);
-        }
-
-        private static T ConvertParameterValue(object parameter)
-        {
-            parameter = parameter is T ? parameter : Convert.ChangeType(parameter, typeof(T));
-            return (T)parameter;
         }
     }
 }
